test(hover): move hover tests to hover_tests.rs
Inline unwrap()/expect() in test code tripped the pre-push guideline check,
which only recognizes test code by file path (tests.rs / a tests/ dir), not
#[cfg(test)] blocks inline in a non-test-suffixed file. Moving to a separate
file matches every other feature's test convention in this codebase
(definition_tests.rs, etc.) and satisfies the check the same way.
